+++ src-plugins/vanilla-model/test/TestOpenSimplexNoise.cpp
@@ -32,7 +32,7 @@ TEST_CASE("3D, Seed 0, Size 10x10, Feature 1.", "[opensimplexnoise]")
         for (int y = 0; y < height; ++y)
         {
             double value = noise.eval(static_cast<double>(x) / feature, static_cast<double>(y) / feature, 0.0);
-            REQUIRE(control.at(x * height + y) == approx(value));
+            REQUIRE(control.at(static_cast<unsigned int>(x * height + y)) == approx(value));
         }
     }
 }
@@ -84,7 +84,7 @@ TEST_CASE("2D, Seed 42, Size 30x20, Feature 1.", "[opensimplexnoise]")
         for (int y = 0; y < height; ++y)
         {
             double value = noise.eval(static_cast<double>(x) / feature, static_cast<double>(y) / feature);
-            REQUIRE(control.at(x * height + y) == approx(value));
+            REQUIRE(control.at(static_cast<unsigned int>(x * height + y)) == approx(value));
         }
     }
 }
@@ -136,7 +136,7 @@ TEST_CASE("3D, Seed 45645, Size 30x20, Feature 10.", "[opensimplexnoise]")
         for (int y = 0; y < height; ++y)
         {
             double value = noise.eval(static_cast<double>(x) / feature, static_cast<double>(y) / feature, 0.0);
-            REQUIRE(control.at(x * height + y) == approx(value));
+            REQUIRE(control.at(static_cast<unsigned int>(x * height + y)) == approx(value));
         }
     }
 }
